+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 ## 1.1.0 under development
 
 - Enh #10: Add `yiisoft/middleware-dispatcher` event listener to keep `Request` always available (@xepozz)
-- Add `Yiisoft/RequestProvider/RequestCookies` request cookie collection object (@hacan359)
+- Add `RequestCookies` class that provides convenient access to request cookies (@hacan359)
 
 ## 1.0.0 March 02, 2024